pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ml-alpha-lab"
version = "0.1.0"
description = "Production-grade ML quantitative trading system"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "lightgbm>=4.0",
    "xgboost>=2.0",
    "torch>=2.0",
    "yfinance>=0.2",
    "mlflow>=2.8",
    "azure-ai-ml>=1.12",
    "azure-identity>=1.15",
    "azure-storage-blob>=12.19",
    "hmmlearn>=0.3",
    "scipy>=1.11",
    "matplotlib>=3.7",
    "seaborn>=0.13",
    "plotly>=5.17",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "click>=8.1",
    "tqdm>=4.66",
    "joblib>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov",
    "ruff>=0.1",
    "mypy>=1.6",
    "black>=23.0",
    "ipykernel",
    "notebook",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 120
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
